Repository
This repository will act as an adapter that delegates to my files where if its time to switch to derby for more realistic data it will come easier that where i don't need to change any from my UI or services. This will be use for my JUnit Testing where i can you fake/in-memory repo.