Update: CourseFiles
I just update and remove some code so that my service test will run smoothly. I got an error because the iterator is being used and what the code is doing is that it is removing from a list that is unmodifiable that returns by s.getEnrolledCourses() that's why i did get a unsupportedoperationexception when i run my serviceTest. but now it fixed