INSERT: Hamcrest and JUnit Test
I inserted hamcrest for my JUnit testing and did a Test class for my enrollment services and change some of the code for my EnrollmentService class. I did some test for the enrolling,  about duplicating checking if the enrollee is already in the system. Checks if a student exceeds its credit limit, it also test dropping courses. it also test if a student is still can enroll and haven't reached its credit limits. I also did a test if a student is enrolled in the first place and last i did a test that checks if they are dropping in a course that they're not enrolled in the first place.