CREATE: CourseSelectionView
This is my main GUI or main frame design. My functions for my program i create and design it as a Tab. I used JTabbedPane class to create a tab in each function this sits in my view. The wiring listeners also sit here in the class which calls the methods when buttons are clicked or pressed. Respectively in each methods.